+++ js/hang/src/publish/video/types.ts
@@ -1,0 +1,28 @@
+export type Source = VideoStreamTrack;
+
+// Stronger typing for the MediaStreamTrack interface.
+export interface VideoStreamTrack extends MediaStreamTrack {
+	kind: "video";
+	clone(): VideoStreamTrack;
+	getSettings(): VideoTrackSettings;
+}
+
+export interface VideoTrackSettings {
+	deviceId: string;
+	groupId: string;
+
+	aspectRatio: number;
+	facingMode: "user" | "environment" | "left" | "right";
+	frameRate: number;
+	height: number;
+	resizeMode: "none" | "crop-and-scale";
+	width: number;
+}
+
+export type VideoConstraints = Omit<
+	MediaTrackConstraints,
+	"autoGainControl" | "channelCount" | "echoCancellation" | "noiseSuppression" | "sampleRate" | "sampleSize"
+> & {
+	// TODO update @types/web
+	resizeMode?: "none" | "crop-and-scale";
+};
