+++ js/hang/src/watch/index.ts
@@ -1,10 +1,9 @@
-export * from "./audio";
+export * as Audio from "./audio";
 export * from "./broadcast";
 export * from "./chat";
 export * from "./location";
 export * from "./preview";
-export * from "./video";
-export * from "./video/detection";
+export * as Video from "./video";
 
 // NOTE: element is not exported from this module
 // You have to import it from @kixelated/hang/watch/element instead.
