+++ js/hang/tsconfig.json
@@ -1,9 +1,8 @@
 {
-	"extends": "../tsconfig.json",
-	"include": ["src"],
-	"exclude": ["node_modules", "dist"],
+	"extends": "../tsconfig.base.json",
 	"compilerOptions": {
 		"outDir": "dist",
 		"rootDir": "./src"
-	}
+	},
+	"include": ["src"]
 }
