+++ js/justfile
@@ -17,6 +17,7 @@ check:
 	pnpm -r run check
 	pnpm -r test
 	pnpm exec biome check
+	pnpm exec madge --circular --extensions ts,tsx .
 	# pnpm exec knip --no-exit-code
 
 # Automatically fix some issues.
