+++ js/moq-clock/tsconfig.json
@@ -1,8 +1,8 @@
 {
-	"extends": "../tsconfig.json",
+	"extends": "../tsconfig.base.json",
 	"compilerOptions": {
 		"rootDir": "./src",
 		"outDir": "./dist"
 	},
-	"include": ["src/**/*"]
+	"include": ["src"]
 }
