+++ js/moq-token/tsconfig.json
@@ -1,8 +1,7 @@
 {
-	"extends": "../tsconfig.json",
-	"include": ["src"],
-	"exclude": ["node_modules", "dist"],
+	"extends": "../tsconfig.base.json",
 	"compilerOptions": {
 		"outDir": "dist"
-	}
+	},
+	"include": ["src"]
 }
