+++ js/package.json
@@ -9,6 +9,7 @@
 		"concurrently": "^9.2.1",
 		"cpy-cli": "^5.0.0",
 		"knip": "^5.63.1",
+		"madge": "^8.0.0",
 		"rimraf": "^6.0.1",
 		"tsx": "^4.20.5",
 		"typescript": "^5.9.2"
